change counter to float

+++ samplers/samplers.go
@@ -76,12 +76,12 @@ type JSONMetric struct {
 type Counter struct {
 	Name  string
 	Tags  []string
-	value int64
+	value float64
 }
 
 // Sample adds a sample to the counter.
 func (c *Counter) Sample(sample float64, sampleRate float32) {
-	c.value += int64(sample) * int64(1/sampleRate)
+	c.value += sample * (1/sampleRate)
 }
 
 // Flush generates a DDMetric from the current state of this Counter.
